+++ src/extension.ts
@@ -5,48 +5,111 @@ import * as vscode from 'vscode';
 import { HC3FS } from './fileSystemProvider';
 import { HC3 } from './hc3';
 
-export function activate(context: vscode.ExtensionContext) {
+let inited = false;
+let O_HC3Console: vscode.LogOutputChannel;
+let O_HC3Events: vscode.LogOutputChannel;
+let O_HC3fslog: vscode.LogOutputChannel;
 
-	console.log('hc3fs says "Hello"');
+export const printHC3ConsoleChannel = (time: string, typ:string, tag:string, message:string): void => {
+	if (typ==='debug') {
+		O_HC3Console.debug(`[${tag}] ${message}`);
+	} else if (typ==='warning') {
+		O_HC3Console.warn(`[${tag}] ${message}`);
+	} else if (typ==='trace') {
+		O_HC3Console.trace(`[${tag}] ${message}`);
+	} else if (typ==='error') {
+		O_HC3Console.error(`[${tag}] ${message}`);
+	}
+};
+
+export const printHC3EventsChannel = (content: string, reveal = false): void => {
+	O_HC3Events.debug(content);
+	if (reveal) {
+		O_HC3Events.show(true);
+	}
+};
+
+export const printHC3fslogChannel = (content: string, reveal = false): void => {
+	O_HC3fslog.debug(content);
+	if (reveal) {
+		O_HC3fslog.show(true);
+	}
+};
+
+export function activate(context: vscode.ExtensionContext) {
+	
+	O_HC3Console = vscode.window.createOutputChannel("HC3 console",{log: true});
+	O_HC3Events = vscode.window.createOutputChannel("HC3 events",{log: true});
+	O_HC3fslog = vscode.window.createOutputChannel("HC3 fslog",{log: true});
+	
+	O_HC3fslog.info('hc3fs says "Hello"');
 	let conf = vscode.workspace.getConfiguration('files');
 	conf.update('autoSave', 'off');
-
+	
 	conf = vscode.workspace.getConfiguration('hc3fs');
-
-	const u = vscode.Uri.parse(conf.url);
 	const hc3 = new HC3(conf);
+	
 	hc3.callHC3("GET", "/settings/info/").then((data) => {
-		console.log("HC3 version: " + data.softVersion);
+		O_HC3fslog.info("HC3 version: " + data.softVersion);
 	}).catch((err) => {
 		vscode.window.showErrorMessage(err);
 	});
+	
 	const hc3Fs = new HC3FS(hc3);
 	context.subscriptions.push(vscode.workspace.registerFileSystemProvider('hc3fs', hc3Fs, { isCaseSensitive: true }));
-
-	// context.subscriptions.push(vscode.commands.registerCommand('hc3fs.reset', _ => {
-	// 	for (const [name] of hc3Fs.readDirectory2(vscode.Uri.parse('hc3fs:/'))) {
-	// 		hc3Fs.delete(vscode.Uri.parse(`hc3fs:/${name}`));
-	// 	}
-	// 	initialized = false;
-	// }));
-
-	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.init', _ => {
-	}));
-
+	inited = true;
+	
 	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.workspaceInit', _ => {
+		if (!inited) {
+			return;
+		}
 		vscode.workspace.updateWorkspaceFolders(0, 0, { uri: vscode.Uri.parse('hc3fs:/'), name: `HC3FS - ${conf.url}` });
 	}));
-
+	
 	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.toggleLog', _ => {
+		if (!inited) {
+			return;
+		}
 		hc3.toggleLog();
+	}));
+	
+	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.logFilter', _ => {
+		if (!inited) {
+			return;
+		}
+		hc3.logFilterPicker();
 	}));
 
 	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.downloadFQA', spec => {
+		if (!inited) {
+			return;
+		}
 		hc3.downloadFQA(spec);
 	}));
-
+	
 	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.downloadScene', spec => {
+		if (!inited) {
+			return;
+		}
 		hc3.downloadScene(spec);
+	}));
+	
+	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.resync', spec => {
+		if (!inited) {
+			return;
+		}
+		if (spec.path === '/QuickApps') {
+			hc3.resyncQA();
+		} else if (spec.path === '/Scenes') {
+			hc3.resyncScenes();
+		}
+	}));
+
+	context.subscriptions.push(vscode.commands.registerCommand('hc3fs.toggleReadOnly', spec => {
+		if (!inited) {
+			return;
+		}
+		hc3.toggleReadOnly(spec);
 	}));
 }